requirements.txt
Flask==2.1.2
numpy==1.22.4
pandas==1.4.2
PyJWT==2.4.0
PyMySQL==1.0.2
gunicorn==20.1.0

